Rename core folder to models
