+++ src/components/map-layer-group/o-map-layer-group.component.ts
@@ -1,6 +1,7 @@
-import { Component, Inject, forwardRef, ViewChildren, ViewEncapsulation } from '@angular/core';
+import { Component, forwardRef, Inject, ViewChildren, ViewEncapsulation } from '@angular/core';
+
 import { OMapComponent, OMapLayerComponent } from '../../components';
-import { LayerConfiguration, LayerGroupConfiguration, OMapLayerGroupsWarehouse } from '../../core';
+import { LayerConfiguration, LayerGroupConfiguration, OMapLayerGroupsWarehouse } from '../../models';
 
 @Component({
   moduleId: module.id,
@@ -21,8 +22,8 @@ import { LayerConfiguration, LayerGroupConfiguration, OMapLayerGroupsWarehouse }
     '[class.o-map-layer-group]': 'true'
   }
 })
-
 export class OMapLayerGroupComponent {
+
   id: string;
   idParent: string;
   name: string;
@@ -106,4 +107,5 @@ export class OMapLayerGroupComponent {
   set description(val: string) {
     this._description = val;
   }
+
 }
