+++ src/components/map-layer/o-map-layer.component.ts
@@ -1,13 +1,14 @@
-import { Component, OnInit, Inject, Injector, EventEmitter, forwardRef, AfterViewInit, ViewEncapsulation } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
-import { Subscription, Observable, combineLatest } from 'rxjs';
-import { map } from 'rxjs/operators';
+import { AfterViewInit, Component, EventEmitter, forwardRef, Inject, Injector, OnInit, ViewEncapsulation } from '@angular/core';
 import * as L from 'leaflet';
-import { MapService } from '../../services';
-import { Center, LayerConfiguration } from '../../core';
-import { Util } from '../../utils';
+import { combineLatest, Observable, Subscription } from 'rxjs';
+import { map } from 'rxjs/operators';
+
 import { OMapComponent, OMapLayerFactory } from '../../components';
 import { ILayerService, OSearchable, OSearchResult } from '../../interfaces';
+import { Center, LayerConfiguration } from '../../models';
+import { MapService } from '../../services';
+import { Util } from '../../utils';
 import { ICRSConfiguration, ICRSConfigurationParameter } from '../map-crs/o-map-crs-configuration.class';
 
 @Component({
@@ -450,4 +451,5 @@ export class OMapLayerComponent implements OnInit, AfterViewInit, OSearchable {
   set menuLabel(val: string) {
     this._menuLabel = val;
   }
+
 }
