+++ src/components/map/o-map-w-search.class.ts
@@ -1,10 +1,11 @@
 import { Observable, Subject } from 'rxjs';
-import { OSearcher, OSearchable } from '../../interfaces';
-import { SearcherDefault } from '../../core';
+
+import { OSearchable, OSearcher } from '../../interfaces';
+import { SearcherDefault } from '../../models';
 import { OMapWBaseLayer } from './o-map-w-baselayer.class';
 
 export class OMapWSearch extends OMapWBaseLayer implements OSearcher {
-	/**
+  /**
 	 * OSearcher implementation
 	 */
   oSearchTitle: string = '';
@@ -27,7 +28,7 @@ export class OMapWSearch extends OMapWBaseLayer implements OSearcher {
     new SearcherDefault('Capas', () => this.mapLayers)
   ];
 
-	/**
+  /**
 	 * OSearcher implementation
 	 */
   search(oSearchValue: string): Observable<Array<OSearchable>> {
@@ -39,4 +40,5 @@ export class OMapWSearch extends OMapWBaseLayer implements OSearcher {
     }));
     return subject.asObservable();
   }
+
 }
