--- src/core.ts
@@ -1,10 +1,0 @@
-export * from './core/BaseLayerCollection.class';
-export * from './core/BaseLayer.class';
-export * from './core/Center.class';
-export * from './core/Feature.class';
-export * from './core/Geometry.class';
-export * from './core/LayerConfiguration.class';
-export * from './core/LayerGroupConfiguration.class';
-export * from './core/LayerGroupsWarehouse.class';
-export * from './core/Location.class';
-export * from './core/Searcher.class';
