+++ src/interfaces/IGeoJSONLayerService.ts
@@ -1,5 +1,5 @@
+import { LayerConfiguration } from '../models';
 import { ILayerService } from './ILayerService';
-import { LayerConfiguration } from '../core/LayerConfiguration.class';
 
 export interface IGeoJSONLayerService extends ILayerService {
   getBaseUrl(layerConf: LayerConfiguration): string;
