+++ src/models/Geometry.class.ts
