+++ src/models/LayerConfiguration.class.ts
