+++ src/models/Location.class.ts
