+++ src/models/Searcher.class.ts
