+++ src/services/MapServiceUtils.ts
@@ -1,6 +1,7 @@
-import { BaseLayerDefault } from '../core';
-import { BaseLayer } from '../interfaces';
 import * as L from 'leaflet';
+
+import { BaseLayer } from '../interfaces';
+import { BaseLayerDefault } from '../models';
 
 const AVAILABLE_STYLE_ATTRIBUTE_KEYS = [
   'stroke', 'color', 'weight', 'opacity', 'lineCap',
@@ -21,7 +22,7 @@ const DEFAULT_BASE_LAYERS = {
     urlTemplate: 'http://server.arcgisonline.com/ArcGIS/rest/services/World_Topo_Map/MapServer/tile/{z}/{y}/{x}',
     options: {
       attribution: 'Tiles &copy; Esri &mdash; Esri, DeLorme, NAVTEQ, TomTom, Intermap, iPC, USGS, FAO, NPS, NRCAN, GeoBase,'
-      + ' Kadaster NL, Ordnance Survey, Esri Japan, METI, Esri China (Hong Kong), and the GIS User Community'
+        + ' Kadaster NL, Ordnance Survey, Esri Japan, METI, Esri China (Hong Kong), and the GIS User Community'
     }
   },
   'CartoDB': {
@@ -29,7 +30,7 @@ const DEFAULT_BASE_LAYERS = {
     urlTemplate: 'http://{s}.basemaps.cartocdn.com/light_all/{z}/{x}/{y}.png',
     options: {
       attribution: '&copy; <a href="http://www.openstreetmap.org/copyright">OpenStreetMap</a> &copy;'
-      + ' <a href="http://cartodb.com/attributions">CartoDB</a>'
+        + ' <a href="http://cartodb.com/attributions">CartoDB</a>'
     }
   }
 };
